static/tests/samesite/fetch.html

<!DOCTYPE html>
<script src="/tests/resources/testharness.js"></script>
<script src="/tests/resources/testharnessreport.js"></script>
<script src="/tests/resources/cookie-helper.js"></script>
<script>
  // Asserts that a request to |origin| contains or does not contain (according
  // to the value of |present|) a cookie named |name| with a value of |value|.
  function assert_http_cookie(origin, name, value, present) {
    return credFetch(origin + "/cookie/list")
        .then(r => r.json())
        .then(obj => {
          assert_cookie(origin, obj, name, value, present);
        });
    }

  function create_test(origin, target, extra, expected, title) {
    promise_test(t => {
      var COOKIE_VALUE = "" + Math.random();
      return create_cookie(origin, HTTP_COOKIE, COOKIE_VALUE, extra)
        .then(_ => {
          return credFetch(target + "/cookie/list")
            .then(r => r.json())
            .then(obj => assert_cookie(origin, obj, HTTP_COOKIE, COOKIE_VALUE, expected));
        });
    }, title);
  }

  // No redirect:
  create_test(ORIGIN, ORIGIN, "SameSite=Lax", true, "Same-host fetches are laxly same-site");
  create_test(ORIGIN, ORIGIN, "SameSite=Strict", true, "Same-host fetches are strictly same-site");
  create_test(SUBDOMAIN_ORIGIN, SUBDOMAIN_ORIGIN, "SameSite=Lax", true, "Subdomain fetches are laxly same-site");
  create_test(SUBDOMAIN_ORIGIN, SUBDOMAIN_ORIGIN, "SameSite=Strict", true, "Subdomain fetches are strictly same-site");
  create_test(CROSS_SITE_ORIGIN, CROSS_SITE_ORIGIN, "SameSite=Lax", false, "Cross-site fetches are not laxly same-site");
  create_test(CROSS_SITE_ORIGIN, CROSS_SITE_ORIGIN, "SameSite=Strict", false, "Subdomain fetches are not strictly same-site");

  // Redirect from {same-host,subdomain,cross-site} to same-host:
  create_test(ORIGIN, redirectTo(ORIGIN, ORIGIN), "SameSite=Lax", true, "Same-host redirecting to same-host fetches are laxly same-site");
  create_test(ORIGIN, redirectTo(ORIGIN, ORIGIN), "SameSite=Strict", true, "Same-host redirecting to same-host fetches are strictly same-site");
  create_test(ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, ORIGIN), "SameSite=Lax", true, "Subdomain redirecting to same-host fetches are laxly same-site");
  create_test(ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, ORIGIN), "SameSite=Strict", true, "Subdomain redirecting to same-host fetches are strictly same-site");
  create_test(ORIGIN, redirectTo(CROSS_SITE_ORIGIN, ORIGIN), "SameSite=Lax", true, "Cross-site redirecting to same-host fetches are laxly same-site");
  create_test(ORIGIN, redirectTo(CROSS_SITE_ORIGIN, ORIGIN), "SameSite=Strict", true, "Cross-site redirecting to same-host fetches are strictly same-site");

  // Redirect from {same-host,subdomain,cross-site} to subdomain:
  create_test(SUBDOMAIN_ORIGIN, redirectTo(ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Lax", true, "Same-host redirecting to subdomain fetches are laxly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Strict", true, "Same-host redirecting to subdomain fetches are strictly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Lax", true, "Subdomain redirecting to subdomain fetches are laxly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Strict", true, "Subdomain redirecting to subdomain fetches are strictly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(CROSS_SITE_ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Lax", true, "Cross-site redirecting to subdomain fetches are laxly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(CROSS_SITE_ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Strict", true, "Cross-site redirecting to subdomain fetches are strictly same-site");

  // Redirect from {same-host,subdomain,cross-site} to cross-site:
  create_test(CROSS_SITE_ORIGIN, redirectTo(ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Lax", false, "Same-host redirecting to cross-site fetches are laxly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Strict", false, "Same-host redirecting to cross-site fetches are not strictly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Lax", false, "Subdomain redirecting to cross-site fetches are laxly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Strict", false, "Subdomain redirecting to cross-site fetches are strictly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(CROSS_SITE_ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Lax", false, "Cross-site redirecting to cross-site fetches are laxly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(CROSS_SITE_ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Strict", false, "Cross-site redirecting to cross-site fetches are strictly same-site");
</script>
